backend/requirements.txt
# --- Core API Framework ---
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic>=2.10.0
pydantic-settings==2.4.0
python-multipart

# --- Utilities ---
numpy<2.0.0
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1

# --- Database Drivers ---
qdrant-client>=1.10.0
redis==5.0.1

# --- AI & LLM Integration ---
openai>=1.50.0
tiktoken>=0.7.0
fastembed>=0.2.0

# --- Document Processing ---
docling>=2.0.0
deepsearch-toolkit
deepsearch-glm

# --- LLM Abstraction & Security ---
litellm>=1.40.0
# [NOWOŚĆ] Biblioteka klienta Langfuse jest wymagana przez LiteLLM do działania callbacków
langfuse>=2.0.0,<3.0.0
presidio-analyzer>=2.2.350
presidio-anonymizer>=2.2.350
spacy>=3.7.0
pl-core-news-md @ https://github.com/explosion/spacy-models/releases/download/pl_core_news_md-3.7.0/pl_core_news_md-3.7.0-py3-none-any.whl
phonenumbers>=8.13.0
arq>=0.25.0

# --- AutoGen ---
autogen==0.7.5
autogen-agentchat==0.7.5

# --- Testing ---
pytest>=8.0.0
pytest-asyncio>=0.23.0